requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
anthropic==0.7.8
python-dotenv==1.0.0
pydantic==2.5.0

